Refactoring & fix changelog
