+++ packages/core-js-compat/src/data.mjs
@@ -1647,6 +1647,8 @@ export const data = {
     firefox: '127',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
+    // Should get next before cloning this
+    // https://bugs.webkit.org/show_bug.cgi?id=289430
     // safari: '18.0', // '17.0',
   },
   'es.set.union.v2': {
@@ -1659,6 +1661,8 @@ export const data = {
     firefox: '127',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
+    // Should get next before cloning this
+    // https://bugs.webkit.org/show_bug.cgi?id=289430
     // safari: '18.0', // '17.0',
   },
   'es.string.at-alternative': {
