+++ packages/core-js/modules/es.set.symmetric-difference.v2.js
@@ -1,10 +1,10 @@
 'use strict';
 var $ = require('../internals/export');
 var symmetricDifference = require('../internals/set-symmetric-difference');
-var setMethodGetNextBeforeCloning = require('../internals/set-method-get-next-before-cloning-detection');
+var setMethodGetKeysBeforeCloning = require('../internals/set-method-get-keys-before-cloning-detection');
 var setMethodAcceptSetLike = require('../internals/set-method-accept-set-like');
 
-var FORCED = !setMethodAcceptSetLike('symmetricDifference') || !setMethodGetNextBeforeCloning('symmetricDifference');
+var FORCED = !setMethodAcceptSetLike('symmetricDifference') || !setMethodGetKeysBeforeCloning('symmetricDifference');
 
 // `Set.prototype.symmetricDifference` method
 // https://tc39.es/ecma262/#sec-set.prototype.symmetricdifference
