+++ packages/core-js/modules/es.set.union.v2.js
@@ -1,10 +1,10 @@
 'use strict';
 var $ = require('../internals/export');
 var union = require('../internals/set-union');
-var setMethodGetNextBeforeCloning = require('../internals/set-method-get-next-before-cloning-detection');
+var setMethodGetKeysBeforeCloning = require('../internals/set-method-get-keys-before-cloning-detection');
 var setMethodAcceptSetLike = require('../internals/set-method-accept-set-like');
 
-var FORCED = !setMethodAcceptSetLike('union') || !setMethodGetNextBeforeCloning('union');
+var FORCED = !setMethodAcceptSetLike('union') || !setMethodGetKeysBeforeCloning('union');
 
 // `Set.prototype.union` method
 // https://tc39.es/ecma262/#sec-set.prototype.union
